requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
docker
streamlit
apache-airflow
requests
